CarIdentifier breaking apart Utilities class
