+++ app/src/main/java/com/tomerpacific/caridentifier/CarManufacturerTranslations.kt
@@ -1,26 +1,5 @@
 package com.tomerpacific.caridentifier
 
-import com.google.mlkit.vision.text.Text
-import com.tomerpacific.caridentifier.model.CarDetails
-import com.tomerpacific.caridentifier.model.CarReview
-
-const val PROS_SECTION_HEBREW = "יתרונות"
-const val CONS_SECTION_HEBREW = "חסרונות"
-const val PROS_SECTION_ENGLISH = "Pros"
-const val CONS_SECTION_ENGLISH = "Cons"
-const val REVIEW_HEBREW = "ביקורת "
-const val REVIEW_ENGLISH = " review"
-
-const val SEVEN_DIGIT_LICENSE_NUMBER_LENGTH_WITH_DASHES = 9
-
-const val EIGHT_DIGIT_LICENSE_NUMBER_LENGTH_WITH_DASHES = 10
-
-private val licensePlateNumberPatterns =
-    listOf(
-        Regex("\\d{2}-\\d{3}-\\d{2}"),
-        Regex("\\d{3}-\\d{2}-\\d{3}"),
-    )
-
 val CAR_MANUFACTURER_NAME_TRANSLATION_TO_ENGLISH =
     mapOf(
         "אאודי" to "Audi",
@@ -140,98 +119,4 @@ val CAR_MANUFACTURER_NAME_TRANSLATION_TO_ENGLISH =
 
 fun getCarManufacturer(manufacturer: String): String {
     return CAR_MANUFACTURER_NAME_TRANSLATION_TO_ENGLISH[manufacturer] ?: "Unknown Manufacturer"
-}
-
-fun formatCarReviewResponse(
-    carReview: String,
-    languageTranslator: LanguageTranslator,
-): CarReview {
-    val prosSection: String
-    val consSection: String
-
-    when (languageTranslator.isHebrewLanguage()) {
-        true -> {
-            prosSection = PROS_SECTION_HEBREW
-            consSection = CONS_SECTION_HEBREW
-        }
-        else -> {
-            prosSection = PROS_SECTION_ENGLISH
-            consSection = CONS_SECTION_ENGLISH
-        }
-    }
-
-    val carReviewLines =
-        carReview
-            .removePrefix("\"")
-            .removeSuffix("\"")
-            .split("\\n")
-    val prosList = mutableListOf<String>()
-    val consList = mutableListOf<String>()
-    var isInProsSection = false
-
-    carReviewLines.forEach { line ->
-        when {
-            line.contains(prosSection, true) -> isInProsSection = true
-            line.contains(consSection, true) -> isInProsSection = false
-            line.isNotBlank() -> if (isInProsSection) prosList.add(line) else consList.add(line)
-        }
-    }
-
-    return CarReview(prosList, consList)
-}
-
-fun concatenateCarMakeAndModel(carDetails: CarDetails): String {
-    val manufacturerName = getCarManufacturer(carDetails.manufacturerName)
-    var commercialName = carDetails.commercialName
-
-    if (doesManufacturerNameExistInCommercialName(manufacturerName, commercialName)) {
-        val indexOfManufacturerName = commercialName.indexOf(manufacturerName, ignoreCase = true)
-        commercialName = commercialName.substring(indexOfManufacturerName + manufacturerName.length).trim()
-    }
-
-    return "$manufacturerName ${commercialName.lowercase().replaceFirstChar { it.titlecase() }
-    } ${
-        carDetails.trimLevel.lowercase().replaceFirstChar { it.titlecase() }
-    } ${carDetails.yearOfProduction}"
-}
-
-fun isLicensePlateNumberValid(licensePlateNumber: String): Boolean {
-    return licensePlateNumberPatterns.any { pattern -> pattern.matches(licensePlateNumber) }
-}
-
-/**
- * Extracts a license plate number from the provided ML Kit Text object.
- *
- * This function sanitizes the text from each line by replacing colons and spaces with dashes.
- * It then finds all lines that match the predefined license plate patterns and returns the one
- * with the highest OCR confidence score.
- *
- * @param text The ML Kit Text object from which to extract the license plate number.
- * @return The license plate number with the highest confidence, or null if none is found.
- */
-fun getLicensePlateNumberFromImageText(text: Text): String? {
-    return text.textBlocks
-        .flatMap { it.lines }
-        .map { line ->
-            Pair(line.text.replace(":", "-").replace(" ", "-"), line)
-        }
-        .filter { (sanitizedText, _) -> isLicensePlateNumberValid(sanitizedText) }
-        .maxByOrNull { (_, line) -> line.confidence }?.first
-}
-
-private fun doesManufacturerNameExistInCommercialName(
-    manufacturerName: String,
-    commercialName: String,
-): Boolean {
-    return commercialName.contains(manufacturerName, ignoreCase = true)
-}
-
-fun handleErrorMessage(exception: Throwable): String {
-    return exception.localizedMessage?.let {
-        if (it.contains("[")) {
-            it.substring(0, it.indexOf("[")).trim()
-        } else {
-            it.trim()
-        }
-    } ?: (exception.message ?: exception.toString())
 }
